Set an attribute of a highlighted vector

Overwrites the conditions, formatters, description, precedence, format_once, init_formatter, or last_formatter attribute of an existing vlighter_highlight vector and validates the result.

Usage

set_conditions(x, value, at = NULL)

set_formatters(x, value, at = NULL)

set_description(x, value, at = NULL)

set_precedence(x, value, at = NULL)

set_format_once(x, value)

set_init_formatter(x, value)

set_last_formatter(x, value)

Arguments

x

[vlightr_highlight]

A highlighted vector to set an attribute of.

value

In set_*, the attribute * that you wish to set.

For example, in set_conditions(), the value must be a function, purrr-style lambda, or list of functions or lambdas to be used as the conditions of x.

See highlight() for the specifics of each attribute.

at

[logical / numeric]

For set_conditions(), set_formatters(), set_description(), and set_precedence(), the indices at which to set the new attribute. Must be a logical or numeric vector.

Value

The vector x with one of it's attributes updated.

See also

Examples

# Let's make these affirmations into exclamations
affirmations <- c("COOL", "Good stuff", "WOW", "Way to go")
exclamations <- highlight(
  affirmations,
  conditions = ~ toupper(.x) == .x,
  formatters = ~ paste(.x, "!")
)
print(affirmations)
#> [1] "COOL"       "Good stuff" "WOW"        "Way to go" 
print(exclamations)
#> <highlight<character>[4]>
#> [1] "COOL !"     "Good stuff" "WOW !"      "Way to go" 
describe_highlight(exclamations)
#> [1] "1:"
#> A conditionally formatted <character> vector with 1 conditional format:
#> 1. Condition 1
#> 
#> Element formatted using:
#> • 1: COOL -> COOL !

# Reset `formatters` to quiet down the exclamation instead
exclamations <- set_formatters(exclamations, ~ tolower(.x))
print(exclamations)
#> <highlight<character>[4]>
#> [1] "cool"       "Good stuff" "wow"        "Way to go" 

# Add a description
exclamations <- set_description(exclamations, "Set Uppercase to Lower")
describe_highlight(exclamations)
#> [1] "1:"
#> A conditionally formatted <character> vector with 1 conditional format:
#> 1. Set Uppercase to Lower
#> 
#> Element formatted using:
#> • 1: COOL -> cool

# The `at` argument is useful with multiple formatters
x_multi <- highlight(
  c(13, 4, 6, 1),
  conditions = list(
    ~ .x %% 2 == 0,
    ~ .x %% 2 != 0,
    ~ .x > 10
   ),
  formatters = list(
    ~ paste(.x, "is even"),
    ~ paste(.x, "is odd"),
    ~ paste(.x, "is big and")
 ),
 last_formatter = wrap("[", "]")
)
print(x_multi)
#> <highlight<double>[4]>
#> [1] [13 is odd is big and] [4 is even]            [6 is even]           
#> [4] [1 is odd]            

# Move the `paste(.x, "is big and")` formatter first
x_multi <- set_precedence(x_multi, value = -1, at = 3)
print(x_multi)
#> <highlight<double>[4]>
#> [1] [13 is big and is odd] [4 is even]            [6 is even]           
#> [4] [1 is odd]