Identify elements of a highlighted vector which meet a condition(s)

condition_id() identifies elements of a highlighted vector x which receive the same conditional formatting. Each element's identifier is a comma separated string of the condition descriptions (attr(x, "description")) corresponding to the conditions (attr(x, "conditions")) met by the element.

Usage

condition_id(x, no_cond = "")

Arguments

x

[vlightr_highlight]

A highlighted vector to identify.

no_cond

[logical(1)]

The identifier assigned to elements which recieve no conditional formatting. Defaults to "".

See also

Examples

# Define a highlight
summary_hl <- highlight_case(
  x = 1:5,
  .x == min(.x, na.rm = TRUE) ~ paste0("(Min = ", .x, ")"),
  .x == quantile(.x, 0.5, na.rm = TRUE) ~ paste0("(Mid = ", .x, ")"),
  .x == max(.x, na.rm = TRUE) ~ paste0("(Max = ", .x, ")"),
  .x > quantile(.x, 0.5, na.rm = TRUE) ~ paste("[ High", .x, "]"),
  .x < quantile(.x, 0.5, na.rm = TRUE) ~ paste("[ Low", .x, "]"),
  .description = c(
    "Min",
    "Middle",
    "Max",
    ">P50",
    "<P50"
  )
)
print(summary_hl)
#> <highlight<integer>[5]>
#> [1] (Min = 1)  [ Low 2 ]  (Mid = 3)  [ High 4 ] (Max = 5) 

# Identify the formatting received by each element
condition_id(summary_hl)
#> [1] "Min"    "<P50"   "Middle" ">P50"   "Max"   

# Identifies element's which meet multiple conditions
summary_hl_multi <- update_highlight(summary_hl, format_once = FALSE)
print(summary_hl_multi)
#> <highlight<integer>[5]>
#> [1] [ Low (Min = 1) ]  [ Low 2 ]          (Mid = 3)          [ High 4 ]        
#> [5] [ High (Max = 5) ]
condition_id(summary_hl_multi)
#> [1] "Min, <P50" "<P50"      "Middle"    ">P50"      "Max, >P50"

# Identify elements which meet no condition
summary_hl[[1]] <- hl(NA)
#> Warning: no non-missing arguments to min; returning Inf
#> Warning: no non-missing arguments to max; returning -Inf
condition_id(summary_hl)
#> [1] ""     "Min"  "<P50" ">P50" "Max" 
condition_id(summary_hl, no_cond = "No Condition")
#> [1] "No Condition" "Min"          "<P50"         ">P50"         "Max"