Subset elements of a highlighted vector which meet a condition(s)

condition_keep() keeps elements of a highlighted vector x which meet the specified conditions (i.e. receive the same conditional formatting).

condition_keep(x, 1) is equivalent to x[attr(x, 'conditions')[[1]](x)].

Usage

condition_keep(x, which = TRUE, meets = c("all", "any", "none"))

Arguments

x

[vlightr_highlight]

A highlighted vector to subset.

which

[character / logical / numeric]

Which conditions are used to subset the vector. Conditions can be selected by index using a numeric or logical vector, or selected using the name of a corresponding description(s) (attr(x, 'description)).

meets

[character(1)]

One of "all", "any", or "none". If "all", then elements of x which meet all of the specified conditions are subset, if "any", then elements which meet "any" of the condition are subset, and if "none", then elements which meet "none" of the conditions are subset

See also

Examples

# Define a highlight
summary_hl <- highlight_case(
  x = 1:5,
  .x == min(.x) ~ paste0("(Min = ", .x, ")"),
  .x == quantile(.x, 0.5) ~ paste0("(Mid = ", .x, ")"),
  .x == max(.x) ~ paste0("(Max = ", .x, ")"),
  .x > quantile(.x, 0.5) ~ paste("[ High", .x, "]"),
  .x < quantile(.x, 0.5) ~ paste("[ Low", .x, "]"),
  .description = c(
    "Min",
    "Middle",
    "Max",
    ">P50",
    "<P50"
  )
)

# Keep the middle element
condition_keep(x, which = 3)
#> Error in eval(expr, envir, enclos): object 'x' not found

# Specify a condition by it's description
condition_keep(x, which = "Middle")
#> Error in eval(expr, envir, enclos): object 'x' not found

# Keep the minimum and maximum elements
condition_keep(x, which = c("Min", "Max"), meets = "any")
#> Error in eval(expr, envir, enclos): object 'x' not found
condition_keep(x, which = c("Min", "Max"), meets = "none")
#> Error in eval(expr, envir, enclos): object 'x' not found
condition_keep(x, which = c("Min", "Max"), meets = "all")
#> Error in eval(expr, envir, enclos): object 'x' not found