Indicate the elements of a highlighted vector which meet a condition(s)

condition_met() returns a logical vector indicating which elements of x meet the specified conditions (i.e. receive the same conditional formatting).

Usage

condition_met(x, which = TRUE, meets = c("all", "any", "none"))

Arguments

x

[vlightr_highlight]

A highlighted vector.

which

[character / logical / numeric]

Which conditions are used to subset the vector. Conditions can be selected by index using a numeric or logical vector, or selected using the name of a corresponding description(s) (attr(x, 'description)).

meets

[character(1)]

One of "all", "any", or "none". The output vector is TRUE for elements of x which:

meet all of the specified conditions, if "all"

meet any of the specified conditions, if "any"

meet none of the specified conditions, if "none"

See also

Examples

# Define a highlight
summary_hl <- highlight_case(
  x = 1:5,
  .x == min(.x) ~ paste0("(Min = ", .x, ")"),
  .x == quantile(.x, 0.5) ~ paste0("(Mid = ", .x, ")"),
  .x == max(.x) ~ paste0("(Max = ", .x, ")"),
  .x > quantile(.x, 0.5) ~ paste("[ High", .x, "]"),
  .x < quantile(.x, 0.5) ~ paste("[ Low", .x, "]"),
  .description = c(
    "Min",
    "Middle",
    "Max",
    ">P50",
    "<P50"
  )
)

# Which element is the minimum?
condition_met(summary_hl, which = 1)
#> [1]  TRUE FALSE FALSE FALSE FALSE

# Specify a condition by it's description
condition_met(summary_hl, which = "Min")
#> [1]  TRUE FALSE FALSE FALSE FALSE

# Which elements are the minimum or maximum?
condition_met(summary_hl, which = c("Min", "Max"), meets = "any")
#> [1]  TRUE FALSE FALSE FALSE  TRUE
condition_met(summary_hl, which = c("Min", "Max"), meets = "none")
#> [1] FALSE  TRUE  TRUE  TRUE FALSE
condition_met(summary_hl, which = c("Min", "Max"), meets = "all")
#> [1] FALSE FALSE FALSE FALSE FALSE