Generate a highlight function with custom default arguments

Produces a partially applied version of the highlight() function, with specified pre-filled arguments. The resulting function can be used to highlight a vector x with a default conditional format.

Assuming that arguments x, conditions, and formatters are valid, the following are equivalent:

highlight(x = x, conditions = conditions, formatters = formatters)

highlighter(conditions = conditions, formatters = formatters)(x)

The output function has the same arguments as highlight(). If arguments conditions, formatters, description, or precedence are supplied to a highlighter function, then the user arguments are appended to the pre-filled arguments prior to highlighting.

If arguments format_once, init_formatter, or last_formatter are supplied to a highlighter, then the supplied arguments override the pre-filled defaults.

Usage

highlighter(
  conditions = list(),
  formatters = list(),
  description = NULL,
  precedence = NULL,
  format_once = FALSE,
  init_formatter = NULL,
  last_formatter = NULL
)

highlighter_case(
  ...,
  .description = NULL,
  .precedence = NULL,
  .format_once = TRUE,
  .init_formatter = NULL,
  .last_formatter = NULL
)

Arguments

conditions

[function / list]

Functions that indicate which elements of x to format. Can be a:

Function, e.g. is.na

A purrr-style lambda, e.g. ~ nchar(.x) > 0, ~ TRUE

A list of functions or lambdas, e.g. list(~ .x < mean(.x), is.infinite)

Each function in conditions will receive x as it's input and must return a logical vector the same length as x or of length 1 (in which case the result will be recycled to the length of x).

Elements of x for which conditions[[i]](x) is TRUE are formatted using the formatter function formatters[[i]]. Consequently, an equal number of conditions and formatters must be supplied.

formatters

[function / list]

Character manipulation functions used to format x. Can be a:

Function, e.g. cli::style_bold

A purrr-style lambda, e.g. ~ paste0(.x, "!"), ~ "Hi"

A list of functions or lambdas, e.g. list(~ cli::col_red, toupper)

When called, a function in formatters will receive a single character vector (of variable length) as it's only argument. A formatter must return a character vector the same length as it's input or of length 1 (in which case the result is recycled to the length of the input character).

ANSI string vectors (class cli_ansi_string) are also supported (see cli::ansi-styles for details).

The same number of formatters and conditions must be supplied.

description, .description

[character / NULL]

An optional description of the format applied by each function in formatters. This information is used by describe_highlight().

If supplied, description must be the same length as formatters and conditions.

precedence, .precedence

[numeric / NULL]

A numeric vector indicating the order in which to apply the formatters. The formatter with the lowest corresponding precedence value is evaluatted first during formatting. By default formatters are applied in the order in which they were supplied.

If supplied, precedence must be the same length as formatters and conditions.

format_once, .format_once

[logical(1)]

A TRUE or FALSE value. Indicates whether an element of x which meets multiple conditions should be formatted only once (using the formatter corresponding to the first condition met) or formatted multiple times (using all of the corresponding formatters).

format_once is FALSE by default in highlight() and TRUE by default in highlight_case().

init_formatter, .init_formatter

[function / NULL]

The first function used to format x. When the highlighted vector is formatted, init_formatter(x) is called prior to conditionally formatting x (i.e. before any of the formatters are called.

If NULL, then format(x) is called instead.

last_formatter, .last_formatter

[function / NULL]

The last function called to format x. The last_formatter is applied after x has been conditionally formatted (i.e. after the init_formatter and formatters functions have been called).

last_formatter will receive a character vector the same length as x as it's only argument. If NULL, the conditionally formatted x is returned as is after formatting.

...

[formula]

This argument replaces the formatters and conditions arguments of highlighter(). The i-th dot supplied is roughly equivalent to conditions[[i]] ~ formatters[[i]].

The left-hand-side and right-hand-side of the formula may be a:

Function, e.g. rlang::is_string, toupper

A purrr-style lambda expression, e.g. TRUE, paste(.x, "?")

The left-hand-side and righ-hand-side of the formula may not be a call to a generator function (i.e. a function which returns another function). Examples include wrap(), color(), and bg(). To use such a function, call it within a lambda expression instead, e.g. color("blue")(.x).

Examples of arguments to ... include:

Colour NA values red: is.na ~ cli::col_red

Add an exclamation mark: toupper(.x) == .x ~ paste0(.x, "!")

Replace 1's with 2's: .x == 1 ~ "2"

Colour the background yellow by default: TRUE ~ bg("yellow")(.x)

Examples

# Mimic an existing highlighted vector
dummy <- c(1L, NA, 0L, 1L, NA)
dummy_hl <- highlight_case(
 x = dummy,
 .x == 0 ~ paste0(.x, "[No]"),
 .x == 1 ~ paste0(.x, "[Yes]"),
 is.na ~ color("red")
)
#> Error in highlight(x = x, conditions = conditions, formatters = formatters,     description = .description, precedence = .precedence, format_once = .format_once,     init_formatter = .init_formatter, last_formatter = .last_formatter): Attempted to create a malformed <vlightr_highlight> vector of type
#> <highlight<integer>>.
#> ✖ Can't evaluate <vlightr_highlight> vector's `format()` method.
#> Caused by error in `format()`:
#> ! Formatter `attr(,"formatters")[[3]]` must produce a <character> vector
#>   of length 1 or `length(x)`.
#> ✖ Produced a <cli_ansi_style/ansi_style> object.
dummy_hlghtr <- highlighter_case(
 .x == 0 ~ paste0(.x, "[No]"),
 .x == 1 ~ paste0(.x, "[Yes]"),
 is.na ~ color("red")
)
print(dummy_hl)
#> Error in eval(expr, envir, enclos): object 'dummy_hl' not found
print(dummy_hlghtr(x = x))
#> Error in eval(expr, envir, enclos): object 'x' not found

# Apply a higlighter to a new vector
dummy_hlghtr(c(-2L, -1L, 0L, 1L, 2L, NA))
#> Error in dummy_hlghtr(c(-2L, -1L, 0L, 1L, 2L, NA)): Attempted to create a malformed <vlightr_highlight> vector of type
#> <highlight<integer>>.
#> ✖ Can't evaluate <vlightr_highlight> vector's `format()` method.
#> Caused by error in `format()`:
#> ! Formatter `attr(,"formatters")[[3]]` must produce a <character> vector
#>   of length 1 or `length(out)`.
#> ✖ Produced a <cli_ansi_style/ansi_style> object.

# Provide additional arguments to a highlighter
dummy_hlghtr(
  x = c(1L, NA, 2L, 0L, 1L, NA, 2L),
  conditions = ~ .x == 2,
  formatters = ~ paste(.x, "[Double Yes]")
)
#> Error in dummy_hlghtr(x = c(1L, NA, 2L, 0L, 1L, NA, 2L), conditions = ~.x ==     2, formatters = ~paste(.x, "[Double Yes]")): Attempted to create a malformed <vlightr_highlight> vector of type
#> <highlight<integer>>.
#> ✖ Can't evaluate <vlightr_highlight> vector's `format()` method.
#> Caused by error in `format()`:
#> ! Formatter `attr(,"formatters")[[3]]` must produce a <character> vector
#>   of length 1 or `length(out)`.
#> ✖ Produced a <cli_ansi_style/ansi_style> object.

# Override default highlighter arguments
secret <- highlighter(
  ~ grepl("^<.*>$", .x),
  ~ strrep("x", nchar(.x) - 2L),
  last_formatter = toupper
)
message <- c("<A>", "Super", "<Secret>", "Message")
secret(message)
#> <highlight<character>[4]>
#> [1] "X"       "SUPER"   "XXXXXX"  "MESSAGE"
secret(message, last_formatter = tolower)
#> <highlight<character>[4]>
#> [1] "x"       "super"   "xxxxxx"  "message"