Restore a highlighted vector's formatting

re_highlight() adds the conditional formatting of highlighted vectors in ... to a highlighted or un-highlighted vector x.

The primary function of re_highlight() is to restore the format method of a highlighted vector after un-highlighting it during an intermediate operation. See the Highlight, Un-Highlight, Re-Highlight section of vignette("vlightr") for more details.

re_highlight(un_highlight(x_hl), x_hl) and x_hl are equivalent for a highlighted vector x_hl.

re_highlight() and rl() are synonyms.

Usage

re_highlight(x, ...)

rl(x, ...)

Arguments

x

[vector]

A vector to highlight with the condition formatting of highlights supplied to .... If x is already a highlighted vector, then it's conditional formatting is combined with that of the highlights supplied to ....

...

[vlighter_highlight]

A highlighted vector used to highlight x.

Examples

# TODO