Update a highlighted vector's format method

update_highlight() sets any of the conditions, formatters, description, precedence, format_once, init_formatter, or last_formatter attributes in an already highlighted vector and validates the result.

Updated attributes must conform to the existing attributes. For example, a supplied list of conditions functions must be the same length as the existing formatters attribute of x.

Usage

update_highlight(
  x,
  conditions,
  formatters,
  description,
  precedence,
  format_once,
  init_formatter,
  last_formatter
)

Arguments

x

[vlightr_highlight]

A highlighted vector to update.

conditions

[function / list]

Functions that indicate which elements of x to format. Can be a:

Function, e.g. is.na

A purrr-style lambda, e.g. ~ nchar(.x) > 0, ~ TRUE

A list of functions or lambdas, e.g. list(~ .x < mean(.x), is.infinite)

Each function in conditions will receive x as it's input and must return a logical vector the same length as x or of length 1 (in which case the result will be recycled to the length of x).

Elements of x for which conditions[[i]](x) is TRUE are formatted using the formatter function formatters[[i]]. If both conditions and formatters are supplied, then they must be the same length.

formatters

[function / list]

Character manipulation functions used to format x. Can be a:

Function, e.g. cli::style_bold

A purrr-style lambda, e.g. ~ paste0(.x, "!"), ~ "Hi"

A list of functions or lambdas, e.g. list(~ cli::col_red, toupper)

When called, a function in formatters will receive a single character vector (of variable length) as it's only argument. A formatter must return a character vector the same length as it's input or of length 1 (in which case the result is recycled to the length of the input character).

ANSI string vectors (class cli_ansi_string) are also supported (see cli::ansi-styles for details).

When formatters is not supplied a default set of formatters are provided. By default, the formatters are a list of cli::bg_br_yellow, cli::bg_br_green, cli::bg_br_red, cli::bg_br_cyan, cli::bg_br_magenta, and cli::bg_br_blue, repeated to the length of conditions.

If the option vlightr.colorful_default_formatters is set via options() to anything other than TRUE, then the default formatters will append numbered annotations "[1]", "[2]", etc. to conditionally formatted elements, instead of using cli::bg_* functions.

description

[character / NULL]

An optional description of the conditional format applied by each function in formatters. This information is used by describe_highlight().

If supplied, description must be the same length as conditions.

precedence

[numeric / NULL]

A numeric vector indicating the order in which to apply the formatters. The formatter with the lowest corresponding precedence value is evaluated first during formatting. By default formatters are applied in the order in which they were supplied.

If supplied, precedence must be the same length as conditions.

format_once

[logical(1)]

A TRUE or FALSE value. Indicates whether an element of x which meets multiple conditions should be formatted only once (using the formatter corresponding to the first condition met) or formatted multiple times (using all of the corresponding formatters).

format_once is FALSE by default in highlight() and TRUE by default in highlight_case().

init_formatter

[function / NULL]

The first function used to format x. When the highlighted vector is formatted, init_formatter(x) is called prior to conditionally formatting x (i.e. before any of the formatters are called.

If NULL, then format(x) is called instead.

last_formatter

[function / NULL]

The last function called to format x. The last_formatter is applied after x has been conditionally formatted (i.e. after the init_formatter and formatters functions have been called).

last_formatter will receive a character vector the same length as x as it's only argument. If NULL, the conditionally formatted x is returned as is after formatting.

See also

Examples

dummies <- c(1L, 0L, NA, 0L, 1L)
dummies_hl <- highlight_case(
  x = dummies,
  .x == 0 ~ paste(.x, "[No]"),
  .x == 1 ~ paste(.x, "[Yes]"),
  TRUE ~ "Don't Know"
)
print(dummies_hl)
#> <highlight<integer>[5]>
#> [1] 1 [Yes]    0 [No]     Don't Know 0 [No]     1 [Yes]   
describe_highlight(dummies_hl)
#> A conditionally formatted <integer> vector with 3 conditional formats:
#> 1. Conditional Format 1
#> 2. Conditional Format 2
#> 3. Conditional Format 3
#> 
#> Element formatted using:
#> • 1:  0 -> 0 [No]
#> • 2:  1 -> 1 [Yes]
#> • 3: NA -> Don't Know

# Add a `description` and `last_formatter`
dummies_hl <- update_highlight(
  x = dummies_hl,
  description = c("Set 0 to No", "Set 1 to Yes", " Otherwise, Don't Know"),
  last_formatter = background("bright_yellow")
)
print(dummies_hl)
#> <highlight<integer>[5]>
#> [1] 1 [Yes]    0 [No]     Don't Know 0 [No]     1 [Yes]   
describe_highlight(dummies_hl)
#> A conditionally formatted <integer> vector with 3 conditional formats:
#> 1. Set 0 to No
#> 2. Set 1 to Yes
#> 3.  Otherwise, Don't Know
#> 
#> Element formatted using:
#> • 1:  0 -> 0 [No]
#> • 2:  1 -> 1 [Yes]
#> • 3: NA -> Don't Know

# Change the `formatters` and remove the `init_formatter`
update_highlight(
  x = dummies_hl,
  formatters = list(~ "No", ~ "Yes", color("red")),
  last_formatter = NULL
)
#> <highlight<integer>[5]>
#> [1] Yes No  NA  No  Yes