Convert a highlighted vector into a highlighter

Convert a highlighted vector x (class vlightr_highlight) into a highlighter function (class vlightr_highligher) which applies the same conditional formatting as that of x.

Usage

as_highlighter(x)

Arguments

x

[vlightr_highlight]

A highlighted vector to convert to a highlighter (class vlightr_highlighter) function. See highlighter() for details.

Value

A highlighter function with the same conditions, formatters, description, precedence, format_once, init_formatter, and last_formatter as x.

Examples

# Highlight a vector
dummy <- highlight_case(
 x = c(1L, NA, 0L, 1L, NA),
 .x == 0 ~ paste0(.x, "[No]"),
 .x == 1 ~ paste0(.x, "[Yes]"),
 is.na ~ cli::col_red,
 .description = c(
    "Responded No",
    "Responded Yes",
    "No Response"
  )
)
print(dummy)
#> <highlight<integer>[5]>
#> [1] 1[Yes] NA     0[No]  1[Yes] NA    
describe_highlight(dummy)
#> A conditionally formatted <integer> vector with 3 conditional formats:
#> 1. Responded No
#> 2. Responded Yes
#> 3. No Response
#> 
#> Element formatted using:
#> • 1:  0 -> 0[No]
#> • 2:  1 -> 1[Yes]
#> • 3: NA -> NA

# Convert the highlighted vector into a highlighter function
dummy_hlghtr <- as_highlighter(dummy)
print(dummy_hlghtr(c(0L, 1L, NA, 2L)))
#> <highlight<integer>[4]>
#> [1] 0[No]  1[Yes] NA     2     
describe_highlight(dummy_hlghtr(c(0L, 1L, NA, 2L)))
#> A conditionally formatted <integer> vector with 3 conditional formats:
#> 1. Responded No
#> 2. Responded Yes
#> 3. No Response
#> 
#> Element formatted using:
#> • 1:  0 -> 0[No]
#> • 2:  1 -> 1[Yes]
#> • 3: NA -> NA