Conditionally format a vector

highlight(), hl(), and highlight_case() create a vlighter_highlight vector with a conditional format() method.

vlighter_highlight is a generic vector superclass which maintains the underlying data of it's subclass while changing how the subclass is formatted and printed. In other words, the vector highlight(1:5) can (with limited legwork) be treated exactly the same as the integer vector 1:5, but has a different format() and print() method.

highlight_case() provides an alternative dplyr::case_when() style syntax to highlight(), but is otherwise similar. Arguments (other than x) to highlight_case() are prefixed with a dot (.).

hl() and highlight() are synonyms.

Usage

highlight(
  x = logical(),
  conditions = list(),
  formatters = list(),
  description = NULL,
  precedence = NULL,
  format_once = FALSE,
  init_formatter = NULL,
  last_formatter = NULL
)

hl(
  x = logical(),
  conditions = list(),
  formatters = list(),
  description = NULL,
  precedence = NULL,
  format_once = FALSE,
  init_formatter = NULL,
  last_formatter = NULL
)

highlight_case(
  x,
  ...,
  .description = NULL,
  .precedence = NULL,
  .format_once = TRUE,
  .init_formatter = NULL,
  .last_formatter = NULL
)

Arguments

x

[vector]

A vector to highlight. Conceptually, a vector is a collection of objects of size 1.

x is considered a vector if:

x is not a data.frame

x is not a bare list

x is otherwise a vector, as defined by vctrs::obj_is_vector()

Atomic vector types "logical", "integer", "double", "complex", "character", and "raw" meet these criteria. As do many common vector classes such as POSIXct, lubridate::interval, or ivs::iv.

conditions

[function / list]

Functions that indicate which elements of x to format. Can be a:

Function, e.g. is.na

A purrr-style lambda, e.g. ~ nchar(.x) > 0, ~ TRUE

A list of functions or lambdas, e.g. list(~ .x < mean(.x), is.infinite)

Each function in conditions will receive x as it's input and must return a logical vector the same length as x or of length 1 (in which case the result will be recycled to the length of x).

Elements of x for which conditions[[i]](x) is TRUE are formatted using the formatter function formatters[[i]]. Consequently, an equal number of conditions and formatters must be supplied.

formatters

[function / list]

Character manipulation functions used to format x. Can be a:

Function, e.g. cli::style_bold

A purrr-style lambda, e.g. ~ paste0(.x, "!"), ~ "Hi"

A list of functions or lambdas, e.g. list(~ cli::col_red, toupper)

When called, a function in formatters will receive a single character vector (of variable length) as it's only argument. A formatter must return a character vector the same length as it's input or of length 1 (in which case the result is recycled to the length of the input character).

ANSI string vectors (class cli_ansi_string) are also supported (see cli::ansi-styles for details).

The same number of formatters and conditions must be supplied.

description, .description

[character / NULL]

An optional description of the format applied by each function in formatters. This information is used by describe_highlight().

If supplied, description must be the same length as formatters and conditions.

precedence, .precedence

[numeric / NULL]

A numeric vector indicating the order in which to apply the formatters. The formatter with the lowest corresponding precedence value is evaluatted first during formatting. By default formatters are applied in the order in which they were supplied.

If supplied, precedence must be the same length as formatters and conditions.

format_once, .format_once

[logical(1)]

A TRUE or FALSE value. Indicates whether an element of x which meets multiple conditions should be formatted only once (using the formatter corresponding to the first condition met) or formatted multiple times (using all of the corresponding formatters).

format_once is FALSE by default in highlight() and TRUE by default in highlight_case().

init_formatter, .init_formatter

[function / NULL]

The first function used to format x. When the highlighted vector is formatted, init_formatter(x) is called prior to conditionally formatting x (i.e. before any of the formatters are called.

If NULL, then format(x) is called instead.

last_formatter, .last_formatter

[function / NULL]

The last function called to format x. The last_formatter is applied after x has been conditionally formatted (i.e. after the init_formatter and formatters functions have been called).

last_formatter will receive a character vector the same length as x as it's only argument. If NULL, the conditionally formatted x is returned as is after formatting.

...

[formula]

For highlighter_case(), a two sided formula with a condition on the left-hand-side and a formatter on the right-hand-side.

This argument replaces the formatters and conditions arguments of highlight(). The i-th dot supplied is roughly equivalent to conditions[[i]] ~ formatters[[i]].

The left-hand-side and right-hand-side of the formula may be a:

Function, e.g. rlang::is_string, toupper

A purrr-style lambda expression, e.g. TRUE, paste(.x, "?")

The left-hand-side and righ-hand-side of the formula may not be a call to a generator function (i.e. a function which returns another function). Examples include wrap(), color(), and bg(). To use such a function, call it within a lambda expression instead, e.g. color("blue")(.x).

Examples of arguments to ... include:

Colour NA values red: is.na ~ cli::col_red

Add an exclamation mark: toupper(.x) == .x ~ paste0(.x, "!")

Replace 1's with 2's: .x == 1 ~ "2"

Colour the background yellow by default: TRUE ~ bg("yellow")(.x)

Value

A highlighted vector (class vlightr_highlight) containing the same data as x.

See also

Examples

# Color NA values red
x <- c(1, 0, NA, 1, 0)
x_hl <- highlight(x, is.na, color("red"))
print(x)
#> [1]  1  0 NA  1  0
print(x_hl)
#> <highlight<double>[5]>
#> [1] 1  0  NA 1  0 
describe_highlight(x_hl)
#> A conditionally formatted <numeric> vector with 1 conditional format:
#> 1. Condition 1
#> 
#> Element formatted using:
#> • 1: NA -> NA

# Label indicators 1 and 0 by adding highlights to `x_hl`
x_hl <- highlight(
  x_hl,
  conditions = list(~ .x == 1, ~ .x == 0),
  formatters = list(~ paste(.x, "[Yes]"), ~ paste(.x, "[No]"))
)
print(x_hl)
#> <highlight<double>[5]>
#> [1] 1 [Yes] 0 [No]  NA      1 [Yes] 0 [No] 
describe_highlight(x_hl)
#> A conditionally formatted <numeric> vector with 2 conditional formats:
#> 1. Condition 1
#> 2. Condition 2
#> 
#> Element formatted using:
#> • 1: 1 -> 1 [Yes]
#> • 2: 0 -> 0 [No]

# Using `dplyr::case_when` style syntax.
# The right hand side `formatters` must be explicit functions.
x_hl_case <- highlight_case(
  x,
  is.na(.x) ~ colour("red"),
  .x == 1 ~ \(x) paste(x, "[Yes]"),
  .x == 0 ~ \(x) paste(x, "[No]"),
  .description = c(
    "Colored Red if NA",
    "Labelled Yes if 1",
    "Labelled No if 0"
  )
)
#> Error in highlight(x = x, conditions = conditions, formatters = formatters,     description = .description, precedence = .precedence, format_once = .format_once,     init_formatter = .init_formatter, last_formatter = .last_formatter): Attempted to create a malformed <vlightr_highlight> vector of type
#> <highlight<double>>.
#> ✖ Can't evaluate <vlightr_highlight> vector's `format()` method.
#> Caused by error in `format()`:
#> ! Formatter `attr(,"formatters")[[1]]` must produce a <character> vector
#>   of length 1 or `length(x)`.
#> ✖ Produced a <cli_ansi_style/ansi_style> object.
print(x_hl_case)
#> Error in eval(expr, envir, enclos): object 'x_hl_case' not found
describe_highlight(x_hl_case)
#> Error in eval(expr, envir, enclos): object 'x_hl_case' not found

# Make a `highlighter` to apply the format of `x_hl_case`
indicator_highlighter <- as_highlighter(x_hl_case)
#> Error in eval(expr, envir, enclos): object 'x_hl_case' not found
indicator_highlighter(c(1, 0, 1, NA, 0, 0))
#> Error in indicator_highlighter(c(1, 0, 1, NA, 0, 0)): could not find function "indicator_highlighter"

# Apply multiple formats to the same element
x_multi <- highlight(
  x = 1:6,
  conditions = list(~ .x %% 2 == 0, ~ .x > 3),
  formatters = list(wrap("<", ">"), wrap("[", "]"))
)
print(x_multi)
#> <highlight<integer>[6]>
#> [1] 1     <2>   3     [<4>] [5]   [<6>]

# Apply a single format to each element with `format_once`
update_highlight(x_multi, format_once = TRUE)
#> <highlight<integer>[6]>
#> [1] 1   <2> 3   <4> [5] <6>

# Use an `init_formatter` to pre-format a highlight
dollar <- highlight_case(
  x = c(10, -1.45, 1.046, -8, NA),
  is.na(.x) ~ \(x) "NA",
  .x > 0 ~ \(x) paste0("-$", x),
  TRUE ~ \(x) paste0("$", x),
  .init_formatter = \(x) sprintf("%.2f", abs(x))
)
print(dollar)
#> <highlight<double>[5]>
#> [1] -$10.00 $1.45   -$1.05  $8.00   NA