Restore a highlighted vector's formatting

re_highlight() adds the conditional formatting of highlighted vectors in ... to a highlighted or un-highlighted vector x.

The primary function of re_highlight() is to restore the format method of a highlighted vector after un-highlighting it during an intermediate operation. See the Highlight, Un-Highlight, Re-Highlight section of vignette("vlightr") for more details.

re_highlight(un_highlight(x_hl), x_hl) and x_hl are equivalent for a highlighted vector x_hl.

re_highlight() and rl() are synonyms.

Usage

re_highlight(x, ...)

rl(x, ...)

Arguments

x

[vector]

A vector to highlight with the condition formatting of highlights supplied to .... If x is already a highlighted vector, then it's conditional formatting is combined with that of the highlights supplied to ....

...

[vlighter_highlight]

A highlighted vector used to highlight x. The conditions, formatters, description, precedence, format_once, init_formatter, and last_formatter of each ... argument are appended to that of x, if x is a highlighted vector, and are supplied to highlight(x, ...) otherwise.

The format_once attribute of the output vector is set to TRUE if format_once is TRUE for any of x or .... Otherwise, format_once is set to FALSE.

The init_formatter (last_formatter) is taken from x, if it is defined (i.e. not NULL), and is otherwise taken from the first argument in ... for which an init_formatter (last_formatter) is defined.

Examples

# Restore a vector's conditional formatting after manipulating it
praise <- c("is cool", "has great vibes", "can do anything")
names <- "David|Aretha|Fats|Marvin|Lauryn"
praise_names <- highlight(
  c("Marvin", "House", "Fats", "Soup"),
  conditions = ~ grepl(names, .x),
  formatters = ~ paste(.x, rep_len(praise, length(.x)))
)
print(praise_names)
#> <highlight<character>[4]>
#> [1] "Marvin is cool"       "House"                "Fats has great vibes"
#> [4] "Soup"                

new_names <- paste(
  un_highlight(praise_names),
  c("Gaye", "MD", "Domino", "Salad")
)
print(new_names)
#> [1] "Marvin Gaye" "House MD"    "Fats Domino" "Soup Salad" 
print(re_highlight(new_names, praise_names))
#> <highlight<character>[4]>
#> [1] "Marvin Gaye is cool"         "House MD"                   
#> [3] "Fats Domino has great vibes" "Soup Salad"                 

# Add several conditional formats to an unhighlighted vector
x <- highlight(1:5, ~ .x > 3, wrap("[", "]"))
y <- highlight(1:5, ~ .x < 2, wrap("<", ">"))
print(x)
#> <highlight<integer>[5]>
#> [1] 1   2   3   [4] [5]
print(y)
#> <highlight<integer>[5]>
#> [1] <1> 2   3   4   5  
print(re_highlight(1:5, x, y))
#> <highlight<integer>[5]>
#> [1] <1> 2   3   [4] [5]

# Or add conditional formats to an already highlighted vector
z <- highlight(1:5, ~ .x %% 2 == 0, wrap("(", ")"))
print(z)
#> <highlight<integer>[5]>
#> [1] 1   (2) 3   (4) 5  
print(re_highlight(z, x, y))
#> <highlight<integer>[5]>
#> [1] <1>   (2)   3     [(4)] [5]  

## Restore the conditional formatting of multiple highlights
positive_adjectives <- "(amazing|great|fantastic)"
positive_phrases <- highlight(
  c("that's amazing", "alright", "you're fantastic"),
  conditions = ~ grepl(positive_adjectives, .x),
  formatters = toupper
)

doubts <- "(but|maybe)"
doubtful_phrases <- highlight(
  c("but I can't go", "I'll be there", "maybe too fantastic"),
  conditions = ~ grepl(doubts, .x),
  formatters = ~ paste0(.x, "...")
)

# Highlights separately
print(positive_phrases)
#> <highlight<character>[3]>
#> [1] "THAT'S AMAZING"   "alright"          "YOU'RE FANTASTIC"
print(doubtful_phrases)
#> <highlight<character>[3]>
#> [1] "but I can't go..."      "I'll be there"          "maybe too fantastic..."

# Highlights combined
paste0(ul(positive_phrases), ", ", ul(doubtful_phrases)) |>
  re_highlight(positive_phrases, doubtful_phrases)
#> <highlight<character>[3]>
#> [1] "THAT'S AMAZING, BUT I CAN'T GO..."       
#> [2] "alright, I'll be there"                  
#> [3] "YOU'RE FANTASTIC, MAYBE TOO FANTASTIC..."